add abstract quoted typed argument type
