+++ src/main/java/com/njdaeger/pdk/command/brigadier/arguments/QuotedStringArgument.java
@@ -2,25 +2,16 @@ package com.njdaeger.pdk.command.brigadier.arguments;
 
 import com.mojang.brigadier.Message;
 import com.mojang.brigadier.StringReader;
-import com.mojang.brigadier.arguments.ArgumentType;
-import com.mojang.brigadier.arguments.StringArgumentType;
-import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.DynamicCommandExceptionType;
-import com.mojang.brigadier.suggestion.Suggestions;
-import com.mojang.brigadier.suggestion.SuggestionsBuilder;
 import com.njdaeger.pdk.command.brigadier.ICommandContext;
-import com.njdaeger.pdk.command.brigadier.CommandContextImpl;
-import io.papermc.paper.command.brigadier.CommandSourceStack;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public class QuotedStringArgument extends BasePdkArgumentType<String, String> {
+public class QuotedStringArgument extends AbstractQuotedTypedArgument<String> {
 
     private static final DynamicCommandExceptionType EMPTY_STRING = new DynamicCommandExceptionType(o -> () -> "Empty strings are not allowed.");
 
@@ -58,36 +49,6 @@ public class QuotedStringArgument extends BasePdkArgumentType<String, String> {
     }
 
     @Override
-    public @NotNull ArgumentType<String> getNativeType() {
-        return StringArgumentType.string();
-    }
-
-    @Override
-    public <S> @NotNull CompletableFuture<Suggestions> listSuggestions(CommandContext<S> context, SuggestionsBuilder builder) {
-        var quotedSuggestions = listSuggestions(new CommandContextImpl((CommandContext<CommandSourceStack>) context));
-        var current = builder.getRemaining();
-        var completingAt = builder.getStart() + (current.contains("\"") ? current.indexOf('"') + 1 : 0);
-        var newBuilder = builder.createOffset(completingAt);
-
-        if (areQuotesBalancedIgnoringEscaped(current) && !current.isBlank()) return newBuilder.buildFuture();
-
-        if (quotedSuggestions.isEmpty()) {
-            if (current.isBlank()) {
-                newBuilder.suggest("\"\"", getDefaultTooltipMessage());
-                return newBuilder.buildFuture();
-            }
-
-            newBuilder.suggest((current.startsWith("\"") ? current.substring(1) : current) + "\"", getDefaultTooltipMessage());
-
-            return newBuilder.buildFuture();
-        }
-
-        quotedSuggestions.forEach((suggestion, tooltip) -> newBuilder.suggest("\"" + suggestion + "\"", tooltip));
-
-        return newBuilder.buildFuture();
-    }
-
-    @Override
     public String convertToCustom(String nativeType, StringReader reader) throws CommandSyntaxException {
         if (!allowEmpty && nativeType.isEmpty()) {
             reader.setCursor(reader.getCursor() - 2);
@@ -99,21 +60,6 @@ public class QuotedStringArgument extends BasePdkArgumentType<String, String> {
     @Override
     public String convertToNative(String s) {
         return s;
-    }
-
-    private static boolean areQuotesBalancedIgnoringEscaped(String str) {
-        int count = 0;
-        boolean escaped = false;
-        for (char c : str.toCharArray()) {
-            if (c == '\\') {
-                escaped = !escaped;
-            } else if (c == '"' && !escaped) {
-                count++;
-            } else {
-                escaped = false;
-            }
-        }
-        return count % 2 == 0;
     }
 
 }
